Login form validation/Http post
Added form validation for login, it's not perfect but it works. Regexp for password could be modified to allow special chars.
Added Http modules and some kind of post function but not sure if it even works, I'll try it later and try to figure how to use the response.