Auth service implemented on login component
HTTP post and login function in auth service. Now implemented in login component. Should figure out how to use the subscription so we can show user success in case of a successfull login and error if login failed. At the moment the boolean that is printed comes before the login event has even happend. I believe it's because the login happens asynchronously.

There is a bug in the password hiding button at the moment, once hiding button is clicked page tries to log user in. No idea why...